# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema de Facturacion Medica API
  version: 0.6.0-draft
  description: |
    Contrato API contract-first en construccion por lotes.
    Lote 1 implementado: auth + base global.
    Lote 2 implementado: patients, doctors, insurance.
    Lote 3 implementado: appointments + medical-records.
    Lote 4 implementado: invoices + payments.
    Lote 5 implementado: catalog.
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /api/v1/invoices/{id}:
    get:
      tags: [Invoices]
      operationId: invoices_get_by_id
      summary: Obtener factura por id
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Factura encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /api/v1/invoices/{id}/view:
    get:
      tags: [Invoices]
      operationId: invoices_get_view_by_id
      summary: Obtener factura por id (vista enriquecida)
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Factura enriquecida encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceViewResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /api/v1/invoices/number/{invoiceNumber}:
    get:
      tags: [Invoices]
      operationId: invoices_get_by_number
      summary: Obtener factura por numero
      parameters:
        - name: invoiceNumber
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Factura encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /api/v1/invoices:
    get:
      tags: [Invoices]
      operationId: invoices_list
      summary: Listar facturas con filtros
      parameters:
        - name: patientId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [DRAFT, PENDING, PARTIAL_PAID, PAID, CANCELLED, OVERDUE]
        - name: startDate
          in: query
          required: false
          schema: { type: string, format: date }
        - name: endDate
          in: query
          required: false
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Lista paginada de facturas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageInvoiceResponse'
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /api/v1/invoices/view:
    get:
      tags: [Invoices]
      operationId: invoices_list_view
      summary: Listar facturas con filtros (vista optimizada)
      parameters:
        - name: patientId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [DRAFT, PENDING, PARTIAL_PAID, PAID, CANCELLED, OVERDUE]
        - name: startDate
          in: query
          required: false
          schema: { type: string, format: date }
        - name: endDate
          in: query
          required: false
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Lista paginada de facturas optimizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageInvoiceListViewResponse'
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
  /api/v1/invoices/{id}/items:
    post:
      tags: [Invoices]
      operationId: invoices_add_item
      summary: Agregar item a factura
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceItemRequest'
      responses:
        '201':
          description: Item agregado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceItemResponse'
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/UnprocessableError' }
  /api/v1/invoices/{id}/insurance-policy:
    patch:
      tags: [Invoices]
      operationId: invoices_assign_insurance_policy
      summary: Asignar o quitar poliza a factura draft
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInsurancePolicyRequest'
      responses:
        '200':
          description: Factura actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/UnprocessableError' }
  /api/v1/invoices/{invoiceId}/items/{itemId}:
    delete:
      tags: [Invoices]
      operationId: invoices_remove_item
      summary: Eliminar item de factura
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Item eliminado
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/UnprocessableError' }
  /api/v1/invoices/{id}/confirm:
    patch:
      tags: [Invoices]
      operationId: invoices_confirm
      summary: Confirmar factura
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Factura actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/UnprocessableError' }
  /api/v1/invoices/{id}/overdue:
    patch:
      tags: [Invoices]
      operationId: invoices_mark_overdue
      summary: Marcar factura como vencida
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Factura actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/UnprocessableError' }
  /api/v1/invoices/{id}/cancel:
    patch:
      tags: [Invoices]
      operationId: invoices_cancel
      summary: Cancelar factura
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Factura actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '422': { $ref: '#/components/responses/UnprocessableError' }
...
  /api/v1/payments/invoice/{invoiceId}:
    get:
      tags: [Payments]
      operationId: payments_get_by_invoice
      summary: Listar pagos por factura
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Lista paginada de pagos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePaymentResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        description: Formato de Spring Data, por ejemplo `lastName,asc`
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
...
    InvoiceItemRequest:
      type: object
      required: [itemType, description, quantity, unitPrice]
      properties:
        serviceId: { type: string, format: uuid, nullable: true }
        medicationId: { type: string, format: uuid, nullable: true }
        itemType:
          type: string
          enum: [service, medication, procedure, other]
        description: { type: string }
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          exclusiveMinimum: 0
    InvoiceItemResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        invoiceId: { type: string, format: uuid }
        serviceId: { type: string, format: uuid, nullable: true }
        serviceName: { type: string, nullable: true }
        medicationId: { type: string, format: uuid, nullable: true }
        medicationName: { type: string, nullable: true }
        itemType:
          type: string
          enum: [service, medication, procedure, other]
        description: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        subtotal: { type: number }
        createdAt: { type: string, format: date-time, nullable: true }
    InvoiceInsurancePolicyRequest:
      type: object
      properties:
        insurancePolicyId: { type: string, format: uuid, nullable: true }
    InvoiceResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        patientId: { type: string, format: uuid }
        patientFirstName: { type: string }
        patientLastName: { type: string }
        appointmentId: { type: string, format: uuid, nullable: true }
        insurancePolicyId: { type: string, format: uuid, nullable: true }
        invoiceNumber: { type: string }
        subtotal: { type: number }
        tax: { type: number }
        total: { type: number }
        insuranceCoverage: { type: number }
        patientResponsibility: { type: number }
        status:
          type: string
          enum: [draft, pending, partial_paid, paid, cancelled, overdue]
        issueDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        notes: { type: string, nullable: true }
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceItemResponse'
        createdAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
    InvoicePatientView:
      type: object
      properties:
        id: { type: string, format: uuid }
        dni: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        allergies: { type: string, nullable: true }
    InvoiceAppointmentView:
      type: object
      properties:
        id: { type: string, format: uuid }
        scheduledAt: { type: string, format: date-time }
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        chiefComplaint: { type: string, nullable: true }
    InvoiceInsurancePolicyView:
      type: object
      properties:
        id: { type: string, format: uuid }
        policyNumber: { type: string }
        providerName: { type: string }
        coveragePercentage: { type: number }
    InvoiceServiceView:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        price: { type: number }
    InvoiceMedicationView:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        requiresPrescription: { type: boolean }
    InvoiceItemView:
      type: object
      properties:
        id: { type: string, format: uuid }
        service:
          allOf:
            - $ref: '#/components/schemas/InvoiceServiceView'
          nullable: true
        medication:
          allOf:
            - $ref: '#/components/schemas/InvoiceMedicationView'
          nullable: true
        itemType:
          type: string
          enum: [service, medication, procedure, other]
        description: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        subtotal: { type: number }
        createdAt: { type: string, format: date-time, nullable: true }
    InvoiceViewResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        invoiceNumber: { type: string }
        patient:
          $ref: '#/components/schemas/InvoicePatientView'
        appointment:
          allOf:
            - $ref: '#/components/schemas/InvoiceAppointmentView'
          nullable: true
        insurancePolicy:
          allOf:
            - $ref: '#/components/schemas/InvoiceInsurancePolicyView'
          nullable: true
        subtotal: { type: number }
        tax: { type: number }
        total: { type: number }
        insuranceCoverage: { type: number }
        patientResponsibility: { type: number }
        status:
          type: string
          enum: [draft, pending, partial_paid, paid, cancelled, overdue]
        issueDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        notes: { type: string, nullable: true }
        items:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceItemView'
        createdAt: { type: string, format: date-time, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
    InvoiceListViewResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        patientId: { type: string, format: uuid }
        patientFirstName: { type: string }
        patientLastName: { type: string }
        invoiceNumber: { type: string }
        total: { type: number }
        patientResponsibility: { type: number }
        status:
          type: string
          enum: [draft, pending, partial_paid, paid, cancelled, overdue]
        issueDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        createdAt: { type: string, format: date-time, nullable: true }
...
    PaymentResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        invoiceId: { type: string, format: uuid }
        invoiceNumber: { type: string }
        amount: { type: number }
        paymentMethod:
          type: string
          enum: [cash, credit_card, debit_card, bank_transfer, check, insurance, other]
        referenceNumber: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        paymentDate: { type: string, format: date-time }
        createdAt: { type: string, format: date-time, nullable: true }
...
    PageInvoiceResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceResponse'
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
    PageInvoiceListViewResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceListViewResponse'
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
    PagePaymentResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PaymentResponse'
        totalElements: { type: integer }
        totalPages: { type: integer }
        size: { type: integer }
        number: { type: integer }
...
  responses:
    BadRequestError:
      description: Solicitud invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Sin permisos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnprocessableError:
      description: Regla de negocio violada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - bearerAuth: []